Move recover_am1 into gcs/constraints/innards/
The recover_am1 helper is a small proof-logging utility that's only
used by the Among and Inverse constraints — no other parts of the
proof framework or the rest of innards consume it. Move it next to
the other constraint-internal helpers (constraints_test_utils.hh,
reified_dispatcher.hh).

Updates the include guard, the source-file list in CMakeLists.txt,
the `#include` paths in among.cc, inverse.cc, and recover_am1.cc
itself, and resorts the include lines in among.cc and inverse.cc so
the new path lands in alphabetical position.
