+++ gcs/constraints/among.cc
@@ -1,9 +1,9 @@
 #include <gcs/constraints/among.hh>
+#include <gcs/constraints/innards/recover_am1.hh>
 #include <gcs/innards/inference_tracker.hh>
 #include <gcs/innards/proofs/names_and_ids_tracker.hh>
 #include <gcs/innards/proofs/proof_logger.hh>
 #include <gcs/innards/proofs/proof_model.hh>
-#include <gcs/innards/proofs/recover_am1.hh>
 #include <gcs/innards/propagators.hh>
 #include <gcs/innards/state.hh>
