+++ gcs/constraints/innards/recover_am1.hh
@@ -1,5 +1,5 @@
-#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INNARDS_PROOFS_RECOVER_AM1_HH
-#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INNARDS_PROOFS_RECOVER_AM1_HH
+#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_INNARDS_RECOVER_AM1_HH
+#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_INNARDS_RECOVER_AM1_HH
 
 #include <gcs/innards/proofs/proof_logger.hh>
