+++ gcs/constraints/inverse.cc
@@ -1,11 +1,11 @@
 #include <gcs/constraints/all_different/gac_all_different.hh>
+#include <gcs/constraints/innards/recover_am1.hh>
 #include <gcs/constraints/inverse.hh>
 #include <gcs/exception.hh>
 #include <gcs/innards/inference_tracker.hh>
 #include <gcs/innards/proofs/names_and_ids_tracker.hh>
 #include <gcs/innards/proofs/proof_logger.hh>
 #include <gcs/innards/proofs/proof_model.hh>
-#include <gcs/innards/proofs/recover_am1.hh>
 #include <gcs/innards/propagators.hh>
 #include <gcs/innards/state.hh>
 #include <gcs/integer.hh>
